+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Attachments page was missing scrollbars.
 - The input field for the alternate name was always overwritten by the original name when loading the location page.
 - Tooltips in the labeling area where not complete and would sometimes overlap with text to be selected.
+- Copy borehole to the correct workgroup if the user only has edit permissions in a single workgroup.
 
 ## v2.1.1116 - 2025-03-17